demo: Move generation into a subcommand in preparation for a verify command
